requirements.txt
PyGithub==2.1.1
pandas==2.1.4
python-dotenv==1.0.0
openai==1.3.7
streamlit==1.29.0
langchain==0.0.350
beautifulsoup4==4.12.2
requests==2.31.0 